fix(codex): keep client bodies out of logs and fuse the label decode
Second review round on #285's implementation.

The `Malformed` arm added last round logged `error = %error`, but
`serde_json::Error`'s Display embeds the offending value: a body that is a
top-level JSON string is echoed in full (verified against serde_json
1.0.151 — a 200-char body produced a 271-char message). That content is
client-controlled and `warn!` becomes a Sentry breadcrumb and an OTel log
record, so the fix for a swallowed error created a log-egress path. Log
`error.line()`, `error.column()`, and `error.classify()` instead — the
sibling `NotAString` arm already avoided this via `json_type_name`, so the
two arms were inconsistent. The zstd `Err` arm keeps its message: those
strings are libzstd-authored constants.

Fuse the decode and the label parse into one bounded unit
(`decode_zstd_and_parse`). Previously `decode_zstd_within` handed the
decoded `Bytes` back to the async caller, which then ran
`serde_json::from_slice` over it on a runtime worker after the admission
permit had already been released — and MAX_DECODE_RATIO lets a ~1 MiB
upload buy a 64 MiB budget, so both the buffer and a multi-millisecond
parse escaped the bound this PR exists to enforce. `extract` now runs in
whichever context the decode ran in, so only its result crosses back.
Streaming the parse off the decoder was considered and rejected:
`decode_within` owns the over-budget-vs-malformed distinction by reading
`budget + 1`, and re-deriving that from `from_reader` would mean inferring
intent from serde's error category. The identity path's synchronous parse is
left alone and marked pre-existing.

Derive INLINE_ZSTD_INPUT_BYTES from the output cap and the measured ratio
(4 KiB -> 2 KiB, with a compile-time assertion tying them). At zstd's real
~12x ratio a 4 KiB input decodes to ~48 KiB, which can never fit the 32 KiB
inline cap, so every body between ~2.7 and 4 KiB was admitted to the inline
probe only to fail it — ~80 us on a worker before redoing the decode
off-thread. The 4 KiB came from INLINE_GZIP_FRAME_BYTES, where gzip's ~4x
ratio makes the band mostly benign. Same provenance error in
`decode_within`'s `with_capacity`, which used gzip's 4x guess for zstd's
~10x output: now a named DECODE_CAPACITY_RATIO.

Correct the compression doc's claim that the ~90 us floor is inherent to
zstd. Measured, a reused `zstd::bulk::Compressor` runs 1 KiB in ~5.8 us vs
~86.8 us for one-shot `encode_all`, so the floor is context setup, not
compression. Not adopted, on cost/benefit rather than feasibility: this runs
once per turn against a multi-second turn, and `Compressor` is not `Sync`,
so it would trade thread-local state for microseconds.

Also split both files' test modules into `<module>/tests.rs`, the pattern
already used seven times under `src/`, bringing `compression.rs` to 320 and
`codex_endpoint.rs` to 486 — both back under the 500-line guideline this
PR had pushed them past. Pure move: 1353 passed / 0 failed / 2 ignored,
unchanged.

Refs #285
